Rename .ios.tsx files to .tsx and simplify wrapper re-exports
