+++ modules/datepicker/datepicker.ts
@@ -1,7 +1,1 @@
-import {Platform} from 'react-native'
-
-import {DatePicker as IosDatePicker} from './datepicker-ios'
-import {DatePicker as AndroidDatePicker} from './datepicker-android'
-
-export const DatePicker =
-	Platform.OS === 'ios' ? IosDatePicker : AndroidDatePicker
+export {DatePicker} from './datepicker-view'
