+++ modules/event-list/event-detail-view.tsx
