+++ source/views/building-hours/detail/link-table-view.tsx
