+++ source/views/building-hours/detail/link-table.ts
@@ -1,6 +1,1 @@
-import {Platform} from 'react-native'
-
-import {LinkTable as LinkTableIos} from './link-table-ios'
-import {LinkTable as LinkTableAndroid} from './link-table-android'
-
-export const LinkTable = Platform.OS === 'ios' ? LinkTableIos : LinkTableAndroid
+export {LinkTable} from './link-table-view'
