+++ source/views/building-hours/detail/schedule-row-view.tsx
