+++ source/views/building-hours/detail/schedule-table-view.tsx
