+++ source/views/building-hours/detail/schedule-table.ts
@@ -1,7 +1,1 @@
-import {Platform} from 'react-native'
-
-import {ScheduleTable as ScheduleTableIos} from './schedule-table-ios'
-import {ScheduleTable as ScheduleTableAndroid} from './schedule-table-android'
-
-export const ScheduleTable =
-	Platform.OS === 'ios' ? ScheduleTableIos : ScheduleTableAndroid
+export {ScheduleTable} from './schedule-table-view'
