+++ source/views/student-orgs/detail-view.tsx
